refactor: BridgeGame에서 InputView를 사용하지 않는 조건 충족
