+++ src/App.js
@@ -1,5 +1,15 @@
+/* eslint-disable class-methods-use-this */
+// const { Console } = require('@woowacourse/mission-utils');
+const BridgeGameControl = require('./BridgeGameControl');
+
 class App {
-  play() {}
+  play() {
+    const bridgeGameControl = new BridgeGameControl();
+    bridgeGameControl.make();
+  }
 }
+
+const app = new App();
+app.play();
 
 module.exports = App;
